distribute simpla-lite package sans IE polyfills
